// src/types/enums.ts
export enum UserRole {
  ADMIN = 'Admin',
  COMPLIANCE_MANAGER_FULL = 'Compliancer Manager FULL',
  COMPLIANCE_MANAGER_READ = 'Compliancer Manager READ',
  COMPLIANCE_MANAGER_WRITE = 'Compliancer Manager WRITE',
  USER = 'User', // Eine Standardrolle für normale Benutzer
  AUDITOR = 'Auditor',
  RISK_MANAGER = 'Risk Manager' // Eine spezifische Rolle für das Risikomanagement
} 